solution updated

+++ Dev1.java
@@ -8,17 +8,17 @@ import java.util.Scanner;
 
 public class Dev1{
       public static void main(String[] args){
-		int num1, num2,num3;
+		long num1, num2,num3;
 		Scanner in = new Scanner(System.in);
 		num1 = in.nextInt();
 
-		for(int i =0; i<num1;i++){
-			num2 = in.nextInt();
-			num3 = in.nextInt();
-			int a[] = new int[num2];
-			for (int j = 0; j < num2; j++){
-         		    if (in.hasNextInt()) {
-                	        a[j] = in.nextInt();
+		for(long i =0; i<num1;i++){
+			num2 = in.nextLong();
+			num3 = in.nextLong();
+			long a[] = new long[(int)num2];
+			for (long j = 0; j < num2; j++){
+         		    if (in.hasNextLong()) {
+                	        a[(int)j] = in.nextLong();
            		    } else {
                 		System.out.println("You didn't provide enough numbers");
                 		break;
@@ -28,12 +28,12 @@ public class Dev1{
 		}
 	}
 
-	static void subArraySum(int arr[], int n, int sum){
-        int curr_sum, i, j;
+	static void subArraySum(long arr[],long n,long sum){
+        long curr_sum, i, j;
         // Pick a starting point
        	 for (i = 0; i < n; i++)
          {
-                curr_sum = arr[i];
+                curr_sum = arr[(int)i];
                 // try all subarrays starting with 'i'
                 for (j = i + 1; j <= n; j++){
                      if (curr_sum == sum){
@@ -42,7 +42,7 @@ public class Dev1{
                      }
                      if (curr_sum > sum || j == n)
                          break;
-                     curr_sum = curr_sum + arr[j];
+                     curr_sum = curr_sum + arr[(int)j];
                 }
 		if(curr_sum==sum)
 			break;
